Unify all exceptions raised
Fixes:
https://github.com/collective/collective.solr/issues/106
