+++ src/collective/solr/indexer.py
@@ -26,7 +26,7 @@ from collective.solr.interfaces import ISolrConnectionManager
 from collective.solr.interfaces import ISolrIndexQueueProcessor
 from collective.solr.interfaces import ICheckIndexable
 from collective.solr.interfaces import ISolrAddHandler
-from collective.solr.solr import SolrException
+from collective.solr.exceptions import SolrConnectionException
 from collective.solr.utils import prepareData
 from socket import error
 from urllib import urlencode
@@ -152,7 +152,7 @@ class BinaryAdder(DefaultAdder):
         try:
             conn.doPost(url, urlencode(postdata, doseq=True), conn.formheaders)
             conn.flush()
-        except SolrException, e:
+        except SolrConnectionException, e:
             logger.warn('Error %s @ %s', e, data['path_string'])
             conn.reset()
 
@@ -221,7 +221,7 @@ class SolrIndexProcessor(object):
                     if adder is None:
                         adder = DefaultAdder(obj)
                     adder(conn, boost_values=boost_values(obj, data), **data)
-                except (SolrException, error):
+                except (SolrConnectionException, error):
                     logger.exception('exception during indexing %r', obj)
 
     def reindex(self, obj, attributes=None):
@@ -261,7 +261,7 @@ class SolrIndexProcessor(object):
             try:
                 logger.debug('unindexing %r (%r)', obj, data)
                 conn.delete(id=data_key)
-            except (SolrException, error):
+            except (SolrConnectionException, error):
                 logger.exception('exception during unindexing %r', obj)
 
     def begin(self):
@@ -282,7 +282,7 @@ class SolrIndexProcessor(object):
                     conn.flush()
                 else:
                     conn.commit(waitSearcher=wait)
-            except (SolrException, error):
+            except (SolrConnectionException, error):
                 logger.exception('exception during commit')
             self.manager.closeConnection()
